Reorganize PSRAM
