+++ esp-hal/src/psram/esp32s2.rs
@@ -1,45 +1,5 @@
-#![cfg_attr(docsrs, procmacros::doc_replace)]
-//! # PSRAM "virtual peripheral" driver (ESP32-S2)
-//!
-//! ## Overview
-//!
-//! The `PSRAM` module provides support for accessing and controlling the
-//! `Pseudo Static Random Access Memory (PSRAM)` on the `ESP32-S2`.
-//!
-//! The `PSRAM` module enables users to interface with the `PSRAM` memory
-//! present on the `ESP32-S2` chip. `PSRAM` provides additional external memory
-//! to supplement the internal memory of the `ESP32-S2`, allowing for increased
-//! storage capacity and improved performance in certain applications.
-//!
-//! //! ## Examples
-//!
-//! ### Quad PSRAM
-//! This example shows how to use PSRAM as heap-memory via esp-alloc.
-//! You need an ESP32S2 with at least 2 MB of PSRAM memory.
-//! Notice that PSRAM example **must** be built in release mode!
-//!
-//! ```rust, ignore
-//! # {before_snippet}
-//! # extern crate alloc;
-//! # use alloc::{string::String, vec::Vec};
-//! #
-//! // Add PSRAM to the heap.
-//! esp_alloc::psram_allocator!(&peripherals.PSRAM, esp_hal::psram);
-//!
-//! init_psram_heap(start, size);
-//!
-//! let mut large_vec: Vec<u32> = Vec::with_capacity(500 * 1024 / 4);
-//!
-//! for i in 0..(500 * 1024 / 4) {
-//!     large_vec.push((i & 0xff) as u32);
-//! }
-//!
-//! let string = String::from("A string allocated in PSRAM");
-//! # {after_snippet}
-//! ```
-
+use super::PsramSize;
 use crate::peripherals::{EXTMEM, SPI0, SPI1};
-pub use crate::soc::psram_common::*;
 
 const EXTMEM_ORIGIN: usize = 0x3f500000;
 
@@ -167,7 +127,7 @@ pub(crate) fn init_psram(config: PsramConfig) {
     }
 
     unsafe {
-        crate::soc::MAPPED_PSRAM.memory_range = EXTMEM_ORIGIN..EXTMEM_ORIGIN + config.size.get();
+        super::MAPPED_PSRAM.memory_range = EXTMEM_ORIGIN..EXTMEM_ORIGIN + config.size.get();
     }
 }
