+++ esp-hal/src/soc/esp32/mod.rs
@@ -7,8 +7,6 @@
 
 crate::unstable_module! {
     pub mod efuse;
-    #[cfg(feature = "psram")]
-    pub mod psram;
     pub mod trng;
 }
 pub mod cpu_control;
